Desktop/sales_insight_automator_RabbitAi/backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
google-generativeai
email-validator
openpyxl
pydantic[email]
python-dotenv
slowapi
